Fix configured amphora agent restarts
In the unlikely even that the amphora agent restarts after being
configured, if the amphora is using systemd, the amphora agent may
fail to start. This is due to a missing "peer ID" command line option.
This patch fixes the systemd unit to have the required "peer ID".

Change-Id: I79cfc9a2e53b249337a189bdb08d13b24807021e
Story: 2002638
Task: 22294